Set callback functions for GLFW
This assumes that Lucid (the object) will always be the window user
pointer.

This will also make a lot more sense, when all attributes of the Lucid
object is public, or at least has all the getters and setters it needs.

This means that all the callback functions will be able to access all
the data as needed.